lint

+++ pacman-core/src/main/java/de/amr/games/pacman/model/common/GameModel.java
@@ -105,7 +105,7 @@ public abstract class GameModel {
 	public static final int EXTRA_LIFE = 10_000;
 
 	// not sure exactly how long Pac-Man is losing power
-	public static final long PAC_POWER_FADING_TICKS = 2 * FPS;
+	public static final int PAC_POWER_FADING_TICKS = 2 * FPS;
 
 	public static final double GHOST_SPEED_HOUSE = 0.5; // unsure
 	public static final double GHOST_SPEED_RETURNING = 2.0; // unsure
